cross-domain-detection/requirements.txt
chainer==4.1.0
chainercv==0.8.0
# cupy==4.1.0
cupy-cuda90==4.1.0
cycler==0.10.0
fastrlock==0.3
filelock==3.0.4
kiwisolver==1.0.1
lxml==4.2.1
matplotlib==2.2.2
numpy==1.14.4
opencv-python==3.4.1.15
Pillow==5.1.0
protobuf==3.5.2.post1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
six==1.11.0
